refactor(#56)!: convert Bound.rangeTo as a member function
