+++ src/commonMain/kotlin/kotools/types/range/NotEmptyRange.kt
@@ -42,15 +42,3 @@ public data class NotEmptyRange<T : Comparable<T>> internal constructor(
         return "$prefix$start;$end$suffix"
     }
 }
-
-/**
- * Creates a [NotEmptyRange] using this bound and the [other] one.
- * If the [other] bound is lower than this bound, the resulting range will
- * [start][NotEmptyRange.start] with the [other] bound.
- */
-@SinceKotoolsTypes("4.2")
-public infix operator fun <T : Comparable<T>> Bound<T>.rangeTo(
-    other: Bound<T>
-): NotEmptyRange<T> =
-    if (value <= other.value) NotEmptyRange(start = this, end = other)
-    else NotEmptyRange(start = other, end = this)
